Revert "We do need the reset config for 32blit"
This reverts commit 1cf97a1350b6bb07727260e7e5dbedc0b124ae35.

We really don't want this. STM32H7 doesn't support it. It causes an
error on startup which is why CLion works "better" with it, but in
fact the debugger failing to start is due to CLion not supporting
the latest version of OpenOCD yet.

OpenOCD change: https://github.com/ntfreak/openocd/commit/7f260f50

CLion bug report: https://youtrack.jetbrains.com/issue/CPP-19052